backend/orchestrator/pyproject.toml
[project]
name = "planner"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "career-database",
    "boto3>=1.40.8",
    "langfuse>=3.3.4",
    "openai-agents[litellm]>=0.3.0",
    "polygon-api-client>=1.15.3",

    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "httpx<0.28.0",
]

[tool.uv.sources]
career-database = { path = "../database", editable = true }
